+++ src/Core/WorkSpace/src/Tabs/GardenTab.tsx
@@ -1,6 +1,6 @@
 import { swap } from '@dbeining/react-atom';
 import { Garden } from '../../../../components/ParkView/Components/Garden';
-import { GardenApi } from '../../../../components/ParkView/Components/VirtualGarden/Container';
+import { GardenApi } from '../../../../components/ParkView/Models/gardenApi';
 import { useFilterApiContext } from '../../../../packages/Filter/Hooks/useFilterApiContext';
 import { WorkspaceFilter } from '../Components/WorkspaceFilter/WorkspaceFilter';
 import { useDataContext } from '../Context/DataProvider';
